Added 'Date' input when creating task
I've added the ability to enter a date when creating a new task. We will use this date to decide when to send a push notification to the user.

Currently, after creating the task, the stored date is not what the user input. Also, the actual push notification itself has not been implemented, nor have I implemented the feature to list the notification under the 'Notifications' tab.

This was only the first step towards the overall goal.